Crate eventually_redis

Redis backend implementation for eventually crate.

Event Store

eventually-redis supports the eventually::EventStore trait through the EventStore type.

Event Subscriber

eventually-redis supports the [eventually::EventSubscriber] trait through the EventSubscriber type.

Structs

Builder

Builder type for EventStore and EventSubscriber types.

BuilderWithSourceName

Second-step builder type for EventStore and EventSubscriber types.

EventStore

Redis backend implementation for eventually::EventStore trait.

EventSubscriber

Redis backend implementation for eventually::EventSubscriber trait.

PersistentSubscription

Subscription implementation with persistent state over a Redis data source.

Constants

STREAM_PAGE_DEFAULT

Default size of a paginated request to Redis XRANGE .. COUNT n for the EventStore::stream and EventStore::stream_all operations.